feat: add envoke unified binary combining renv and kctx
- Add cmd/envoke — single binary entry point for both tools
  - envoke renv <cmd>       → full renv command tree
  - envoke kctx <cmd>       → full kctx command tree
  - envoke resolve [file]   → combined .env resolver (secrets + kubeconfigs)
  - envoke shell-init       → combined bash/zsh/fish shell integration
  - envoke clear-cache      → clears both renv and kctx caches
  - envoke watch            → single watcher for both sleep/lock events

- Extract internal/cli/renv and internal/cli/kctx packages
  - cmd/renv and cmd/kctx become thin wrappers (~13 lines each)
  - All command logic lives in internal/cli/{renv,kctx}/root.go

- Add KCTX_ .env directive support in envoke resolve
  - KCTX_PROD=bw://kubernetes/prod loads kubeconfig 'prod' into kctx named store
  - KCTX_STAGING=vault://secret/k8s loads kubeconfig 'staging'
  - Names validated before any secret fetch (fail-fast, no wasted prompts)
  - Regular env secrets (bw://, vault://) resolve as before

- envoke shell-init emits renv() and kctx() shell functions that call
  envoke renv / envoke kctx — only the envoke binary needs to be in PATH.
  A single combined watcher replaces two separate watcher processes.

- Update Makefile: build-envoke target, install includes envoke
- Update flake.nix: envoke package, nix run now defaults to envoke
