+++ cmd/envoke/detach_windows.go
@@ -1,0 +1,5 @@
+//go:build windows
+
+package main
+
+func detachFromTerminal() {}
