+++ internal/cli/renv/detach_windows.go
@@ -1,0 +1,5 @@
+//go:build windows
+
+package renv
+
+func detachFromTerminal() {}
